+++ core/enums/permissions.ts
@@ -1,0 +1,5 @@
+export enum PermissionLevels {
+  Admin,
+  ServerOwner,
+  BotOwner,
+}