+++ docs/src/misc/changelog.md
@@ -6,15 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.1.0 - 2022-03-19
+
 Initial release of AkumaKodo
 
 ### Added
+
 - Interaction Command handling
 - Lots of util functions
 
 ### Changed
+
 - Added internal slash command handling using events. You can now enable or disable this.
 - Logger Errors to warn on un-stable modes.
 
 ### Removed
-- For now until slash commands are fully implemented, I will not focus on message based commands. You will have to implement them yourself for now.
+
+- For now, until slash commands are fully implemented, I will not focus on message-based commands. You will have to implement them yourself.